finishing work for getting a working server_key
This involves creating a new key, or getting existing one from local server and checking if it still works with global server if not obtaining a new working key and replacing the old with the new one, etc. Still to verify if cyclical getting the working server_key works or possibly to do it from scratch and removing the consequences of getting the working key much later (internet connection available after an hour f.e.).