Completing global server sync for create and update
Now if app is inited there is f.e a ConditionModelUser if it creates  ConditionModelContact model - it is stored locally and synchronized on 'global server' sometimes called global server aspect of any server - each ConditionModelData object has capability to be local server (local app storage) AND global server at the same time. Also when you update a property - f.e. someModel['description'] = 'abcd'; it is updated locally and on global server. It should work now it will be improved and a http proxy ConditionDataManagement driver (for client and for server side) are to be done some time next.