2, 3 Months of changes + resulting regression fixes
No will to remind. In short the idea was implemented that this entire library models, including app model are managed, some more independent standalone apps are allowed. No different object but "twin" settings app model, typical model, ConditionDataManagement driver object can exist in the app which is going to be implemented further (operator == and hashCode is overwritten for this and related classes read further). ConditionDataManagement got to be unfinished interface and while probably it doesn't require this but an sql driver implementation now is divided to three classes, the class itself that now has a db wrapper which peforms sql queries to the db (js, wasm, c++ whatever)  but in a standarized way, the third class is settings class for settings object inside a wrapper. So These combined 3 type of objects let's you know if a drive is duplicate. If there is duplicate model, driver, app model an exception is going to be thrown with the already existing the same object in exception class.